Admin role 10. Ability to change tenants. Clean up old code. Ensure that only data from same tenant as the user tenant are fetched or changed or deleted by user. And more.

+++ public/controllers/tenant.controller.js
@@ -1,7 +1,7 @@
 angular.module('familielejr')
 
-.controller('tenantCtrl', ['$scope', '$http', '$location', '$route', '$window', 'AuthService', 'YearService', 
-function($scope, $http, $location, $route, $window, AuthService, YearService) {
+.controller('tenantsCtrl', ['$scope', '$http', '$location', '$route', 'AuthService', 
+function($scope, $http, $location, $route, AuthService) {
 
     // console.log("In the tenant controller")
     $scope.isLoggedIn = false;
@@ -9,31 +9,35 @@ function($scope, $http, $location, $route, $window, AuthService, YearService) {
         if (AuthService.isLoggedIn()) {
             $scope.isLoggedIn = true;
             $scope.role = AuthService.userRole();
+            if ($scope.role == 10) {
+                $http({
+                    method: 'GET',
+                    url: '/tenants',
+                    headers: {
+                        'x-auth': localStorage.userToken
+                    }
+                }).then(function(response) {
+                    // console.log(`Success. Status: ${response.status}`);
+                    $scope.tenants = response.data;
+                    // console.log($scope.tenants[0]);
+                }, function errorCallback(response) {
+                    console.log(`Error. Status: ${response.status}`);
+                });
+            } else {
+                console.log(`You are NOT the Big ADMIN, so you do not have access here!!`);
+            };
         };
     });
 
     $scope.newTenant = false;
     $scope.editTenant = false;
 
-    $http({
-        method: 'GET',
-        url: '/tenants',
-        headers: {
-            'x-auth': localStorage.userToken
-        }
-    }).then(function(response) {
-        // console.log(`Success. Status: ${response.status}`);
-        $scope.tenants = response.data;
-        // console.log($scope.tenants[0]);
-    }, function errorCallback(response) {
-        console.log(`Error. Status: ${response.status}`);
-    });
-
     $scope.newTenantToggle = function() {
         if ($scope.newTenant) {
             $scope.newTenant = false;
         } else {
             $scope.newTenant = true;
+            $scope.startyear = new Date().getFullYear();
         };
     };
 
@@ -53,7 +57,8 @@ function($scope, $http, $location, $route, $window, AuthService, YearService) {
         var tenant = {
             tenantName: $scope.tenantname,
             description: $scope.description,
-            startYear: (new Date()).getFullYear(),
+            // startYear: (new Date()).getFullYear(),
+            startYear: $scope.startyear,
             subscriptions: subscriptions
         };
 
@@ -102,20 +107,106 @@ function($scope, $http, $location, $route, $window, AuthService, YearService) {
         });
     };
 
-    $scope.editTenantToggle = function(tenant) {
+}])
+
+.controller('tenantdetailsCtrl', ['$scope', '$http', '$location', '$routeParams', '$route', '$window', 'AuthService', 
+function($scope, $http, $location, $routeParams, $route, $window, AuthService) {
+    
+    // console.log("In the tenant details controller")
+    $scope.isLoggedIn = false;
+    AuthService.getUserStatus().then(function() {
+        if (AuthService.isLoggedIn()) {
+            $scope.isLoggedIn = true;
+            $scope.role = AuthService.userRole();
+        };
+    });
+
+    // $scope.newTenant = false;
+    $scope.editTenant = false;
+    $scope.tenantDeletable = true;
+
+    $http({
+        method: 'GET',
+        url: '/tenants/' + $routeParams.id,
+        headers: {
+            'x-auth': localStorage.userToken
+        }
+    }).then(function(tenant) {
+        // console.log(`Success. Tenant Status: ${tenant.status}`);
+        $scope.tenant = tenant.data.tenant;
+        // console.log($scope.tenant.tenantName);
+        return $http({
+            method: 'GET',
+            url: 'events/tenant/' + $routeParams.id,
+            headers: {
+                'x-auth': localStorage.userToken
+            }
+        });
+    }).then(function(events) {
+        // console.log(`Success. Events Status: ${events.status}`);
+        $scope.events = events.data;
+        if ($scope.events.length > 0) {
+            // console.log(`Name of first event: ${$scope.events[0].eventName}`);
+            $scope.tenantDeletable = false;
+        };
+        return $http({
+            method: 'GET',
+            url: 'docs/tenant/' + $routeParams.id,
+            headers: {
+                'x-auth': localStorage.userToken
+            }
+        });
+    }).then(function(docs) {
+        // console.log(`Success. Docs Status: ${docs.status}`);
+        $scope.docs = docs.data;
+        if ($scope.docs.length > 0) {
+            // console.log(`Name of first doc: ${$scope.docs[0].filename}`);
+            $scope.tenantDeletable = false;
+        };
+        return $http({
+            method: 'GET',
+            url: 'users/tenant/' + $routeParams.id,
+            headers: {
+                'x-auth': localStorage.userToken
+            }
+        });
+    }).then(function(users) {
+        // console.log(`Success. Users Status: ${users.status}`);
+        $scope.users = users.data;
+        if ($scope.users.length > 0) {
+            // console.log(`Email of first user: ${$scope.users[0].email}`);
+            $scope.tenantDeletable = false;
+        };
+
+        return $http({
+            method: 'GET',
+            url: 'users/user/' + $scope.tenant._admin,
+            headers: {
+                'x-auth': localStorage.userToken
+            }
+        });
+    }).then(function(admin) {
+        // console.log(`Success. Users Status: ${users.status}`);
+        $scope.admin = admin.data;
+        console.log(`Admin: ${$scope.admin.name.firstname}`);
+    }, function errorCallback(response) {
+        console.log(`Error. Status: ${response.status}`);
+    });
+
+    $scope.editTenantToggle = function() {
+        // console.log(`In editTenantToggle`);
         if ($scope.editTenant) {
             $scope.editTenant = false;
         } else {
-            $scope.tenantToEdit = tenant;
-            $scope.editTenantname = tenant.tenantName;
-            $scope.editDescription = tenant.description;
-            $scope.editID = tenant._id;
+            // console.log(`1. Edit Tenant: ${$scope.tenant.tenantName}`);
+            $scope.tenantToEdit = $scope.tenant;
+            $scope.editID = $scope.tenant._id;
             $scope.editTenant = true;
         };
     };
 
     $scope.tenantEdit = function() {
-        console.log(`Name: ${$scope.tenantToEdit.tenantName}, description: ${$scope.tenantToEdit.description}`)
+        console.log(`Name: ${$scope.tenantToEdit.tenantName}, startYear: ${$scope.tenantToEdit.startYear} description: ${$scope.tenantToEdit.description}`)
         var data = $scope.tenantToEdit;
 
         $http({
@@ -135,7 +226,7 @@ function($scope, $http, $location, $route, $window, AuthService, YearService) {
     };
 
     $scope.removeTenant = function(tenant) {
-        if ($window.confirm('Bekræft venligst at du vil slette tenant '+tenant.tenantname)) {
+        if ($scope.tenantDeletable && $window.confirm('Bekræft venligst at du vil slette tenant '+tenant.tenantname)) {
             $http({
                 method: 'DELETE',
                 url: 'abouts/tenant/'+tenant._id,
@@ -163,3 +254,4 @@ function($scope, $http, $location, $route, $window, AuthService, YearService) {
     };
 
 }]);
+
